in the bindings, changed ClusterTransformer to ForkTransormer for gelu + some debug prints

+++ Deeploy/Targets/PULPOpen/Templates/RQSiGELUTemplate.py
@@ -46,8 +46,8 @@ class _PULPRQSiGELUTemplate(NodeTemplate):
 referenceTemplate = _PULPRQSiGELUTemplate("""
 // Requantized iGELU parallelized for PULP (Name: ${nodeName}, Op: ${nodeOp})
 
-u_int32_t core_id         = pi_core_id();
-u_int32_t log2_core       = log2(NUM_CORES);
+int8_t core_id         = pi_core_id();
+int8_t log2_core       = log2(NUM_CORES);
 int16_t chunk       = (${size} >> log2_core) + (((${size}) & (NUM_CORES - 1)) != 0);
 int16_t chunk_start = MIN(chunk * core_id, ${size});
 int16_t chunk_stop  = MIN(chunk_start + chunk, ${size});
@@ -68,6 +68,8 @@ if (chunk_start < chunk_stop) {
         &${shift_scalar}[0]          // scalar shift pointer
     );
 }
+
+// pi_cl_team_barrier();
 
 """)
